requirements.txt
python-dotenv
langchain_community
pypdf
wikipedia
langchain-text-splitters
langchain-google-genai
langchain-ollama
langchain_huggingface
sentence_transformers
torch
streamlit
langchain_groq
arxiv
faiss-cpu
validators
unstructured
duckduckgo-search
Flask
pillow
numpy